fix bug in scoring edge-case
when a character is in the word in a different position, it should only
be marked as 'incorrect position' when the character at the other
position is also incorrect.

i.e. if the word is 'erand' but the guess is 'eerie' then the second and
third 'e' in the guess should be marked as 'incorrect' rather than
'incorrect position'